+++ lib/double/double.dart
@@ -52,16 +52,21 @@ extension Double_ on double {
     return Tween<double>(begin: this, end: end);
   }
 
+  @Deprecated("Use isBetween() instead")
+  bool between(num first, num second) {
+    return this.isBetween(first, second);
+  }
+
   /// Returns a [bool] if [this] value is between (including) the two
   /// numeric values [first] and [second].
   ///
   /// Example:
   /// ```dart
-  /// 100.0.between(50, 150) // true;
-  /// 100.0.between(50.0, 150.0) // true;
-  /// 100.0.between(100.0, 100.0) // true;
+  /// 100.0.isBetween(50, 150) // true;
+  /// 100.0.isBetween(50.0, 150.0) // true;
+  /// 100.0.isBetween(100.0, 100.0) // true;
   /// ```
-  bool between(num first, num second) {
+  bool isBetween(num first, num second) {
     ArgumentError.checkNotNull(first, "first");
     ArgumentError.checkNotNull(second, "second");
     final lower = min(first, second);